scripts/clipcap/pyproject.toml
[project]
name = "clipcap-script"
version = "0.1.0"
requires-python = "~=3.10.0"
dependencies = [
    "torch",
    "requests",
    "pillow",
    "transformers",
    "clipcap @ git+https://github.com/michelecafagna26/CLIPCap.git",
    "clip @ git+https://github.com/openai/CLIP.git",
] 